refactor: domain package추가
