+++ src/main/java/lotto/domain/LottoException.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public class LottoException extends IllegalArgumentException {
