+++ src/main/java/lotto/domain/LottoNumbersSupplier.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 @FunctionalInterface
 public interface LottoNumbersSupplier {
