+++ src/test/java/lotto/LottoNumbersTest.java
@@ -8,6 +8,10 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
+import lotto.domain.LottoNumber;
+import lotto.domain.LottoNumberMatchCount;
+import lotto.domain.LottoNumbers;
+import lotto.domain.RandomUtils;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
