+++ src/test/java/lotto/LottoResultStatisticsTest.java
@@ -5,6 +5,9 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.Arrays;
 import java.util.List;
+import lotto.domain.LottoNumberMatchCount;
+import lotto.domain.LottoResult;
+import lotto.domain.LottoResultStatistics;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
